requirements.txt
python-telegram-bot
python-dotenv
python-telegram-bot[callback-data]